Fixes to item resize functionality. For #94.

+++ model/variable.cc
@@ -467,7 +467,7 @@ void VariableBase::draw(cairo_t *cairo) const
   float w, h, hoffs, scaleFactor;
   w=rv.width()*z; 
   h=rv.height()*z;
-  scaleFactor=max(1.0,min(static_cast<double>(iWidth())*z/w,static_cast<double>(iHeight())*z/h));
+  scaleFactor=max(1.0f,min(iWidth()*z/w,iHeight()*z/h));
   if (rv.width()<0.5*iWidth()) w=0.5*iWidth()*z;
   if (rv.height()<0.5*iHeight()) h=0.5*iHeight()*z;
   rv.setFontSize(12*scaleFactor*z);
